HHH-17345 - Adjust mapping XSD and JAXB
https://hibernate.atlassian.net/browse/HHH-17345
